Refactored GenericValue
